// src/app/Models/InmueblesModels.ts
export class InmueblesModels {

    MatriculaInmobiliariaInmueble: string;
    ChipInmueble: string;
    TipoInmueble: string;
    NomenclaturaInmueble: string;
    AreaPrivadaInmueble: number;
    AreaConstruidaInmueble: number;
    NumeroEscrituraInmueble: string;
    AlcobasInmueble: number;
    BanosInmueble: number;
    VehiculoInmueble: number;
    IdLocativaInmueble: number;
    CedulaPropietarioInmueble: number;
    MatriculaInmobiliariaProyectoInmueble: number;
}